pyproject.toml
[tool.poetry]
package-mode = false  

[tool.poetry.dependencies]
python = "^3.13"
sqlalchemy = "^2.0.40"
nats-py = "^2.10.0"
environs = "^14.1.1"
psycopg2-binary = "^2.9.10"
